pyproject.toml
[tool.poetry]
name = "fastapi-user"
version = "0.1.0"
description = "user domain"
authors = ["LimdaeIl <optional94@naver.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = "^0.129.0"
uvicorn = "^0.41.0"
sqlalchemy = "^2.0.46"
alembic = "^1.18.4"
pymysql = "^1.1.2"
pydantic-settings = "^2.13.0"
redis = "^7.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.5.0"
cryptography = "^46.0.5"
pydantic = {extras = ["email"], version = "^2.12.5"}
bcrypt = "4.1.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
